+++ roseau/load_flow/exceptions.py
@@ -61,6 +61,7 @@ class RoseauLoadFlowExceptionCode(StrEnum):
 
     # Transformer and line
     BAD_MAX_LOADING_VALUE = auto()
+    BAD_PARAMETERS_ID = auto()
 
     # Switch
     BAD_SWITCH_ID = auto()
@@ -121,8 +122,6 @@ class RoseauLoadFlowExceptionCode(StrEnum):
     DGS_MISSING_REQUIRED_DATA = auto()
 
     # JSON export
-    JSON_LINE_PARAMETERS_DUPLICATES = auto()
-    JSON_TRANSFORMER_PARAMETERS_DUPLICATES = auto()
     JSON_PREF_INVALID = auto()
     JSON_NO_RESULTS = auto()